@@ -1,6 +1,6 @@
 {
   "name": "react-schema-form-example",
-  "version": "0.4.1",
+  "version": "0.3.8",
   "description": "React json schema form example",
   "keywords": [
     "json schema",
@@ -24,34 +24,31 @@
     "url": "https://github.com/networknt/react-schema-form.git"
   },
   "dependencies": {
-    "babel-cli": "^6.24.1",
-    "babel-core": "^6.5.2",
-    "babel-eslint": "^7.1.0",
-    "babel-loader": "^6.2.7",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.16.0",
-    "babel-preset-stage-0": "^6.16.0",
-    "babel-runtime": "^6.18.0",
-    "brace": "^0.9.0",
-    "css-loader": "^0.25.0",
-    "eslint": "^3.9.1",
-    "eslint-plugin-react": "^6.5.0",
-    "jquery": "^3.1.1",
-    "less": "^2.7.1",
-    "less-loader": "^2.2.3",
+    "@material-ui/core": "^1.4.3",
+    "babel-eslint": "^8.2.6",
+    "css-loader": "^1.0.0",
+    "install": "^0.12.1",
+    "jquery": "^3.3.1",
     "lodash": "^4.17.10",
-    "material-ui": "^1.0.0-beta.45",
-    "react": "^16.3.2",
-    "react-ace": "^6.1.0",
-    "react-dom": "^16.3.2",
-    "react-hot-loader": "^3.0.0-beta.6",
+    "npm": "^6.3.0",
+    "react": "^16.4.2",
+    "react-ace": "^6.1.4",
+    "react-dom": "^16.4.2",
     "react-schema-form": "file:../",
-    "react-select": "^1.0.0-rc.5",
-    "style-loader": "^0.13.1",
-    "webpack": "^2.7.0",
-    "webpack-dev-server": "^2.6.1"
+    "react-select": "^2.0.0",
+    "style-loader": "^0.21.0",
+    "supports-color": "^5.4.0",
+    "webpack-cli": "^3.1.0"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "@webpack-cli/migrate": "^0.1.0",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.3",
+    "babel-loader": "^7.1.5",
+    "babel-preset-env": "^1.7.0",
+    "webpack": "^4.16.5",
+    "webpack-dev-server": "^3.1.5"
+  },
   "scripts": {
     "start": "webpack-dev-server -d --config webpack.config.js"
   }
