@@ -1,195 +1,173 @@
-/**
- * Created by steve on 12/09/15.
- */
-import React, { Component } from "react";
-import { utils } from "react-schema-form";
-import { SchemaForm } from "react-schema-form";
-require("react-select/less/select.less");
-import Select from "react-select";
-var $ = require("jquery");
-import AceEditor from "react-ace";
-require("brace/ext/language_tools");
-require("brace/mode/json");
-require("brace/theme/github");
-import Button from "material-ui/Button";
+import React from 'react'
+import {SchemaForm, utils} from 'react-schema-form'
+import AceEditor from 'react-ace'
 
-class ExamplePage extends Component {
+// RcSelect is still in migrating process so it's excluded for now
+// import RcSelect from 'react-schema-form-rc-select/lib/RcSelect';
+import Button from '@material-ui/core/Button'
+
+class ExamplePage extends React.Component {
+
+    tempModel = {
+        'comments': [
+            {
+                'name': '1'
+            },
+            {
+                'name': '2'
+            }
+        ]
+    }
+
     state = {
         tests: [
-            { label: "Simple", value: "data/simple.json" },
-            { label: "Simple Array", value: "data/simplearray.json" },
-            { label: "Basic JSON Schema Type", value: "data/types.json" },
-            { label: "Basic Radios", value: "data/radio.json" },
-            { label: "Condition", value: "data/condition.json" },
-            { label: "Kitchen Sink", value: "data/kitchenSink.json" },
-            { label: "Login", value: "data/login.json" },
-            { label: "Date", value: "data/date.json" },
-            { label: "DateTime", value: "data/datetime.json" },
-            { label: "Readonly", value: "data/readonly.json" },
-            { label: "Array", value: "data/array.json" },
-            { label: "Object", value: "data/object.json" },
-            { label: "ArraySelect", value: "data/arrayselect.json" }
+            {label: "provided for test 1", value: 'data/Token.json'},
+            {label: "provided for test 2", value: 'data/TokenListing.json'},
+            {label: "provided for test 3", value: 'data/TokenProject.json'},
+            {label: "provided for test 4", value: 'data/TokenProjectLegalEntities.json'},
+            {label: "Simple", value: 'data/simple.json'},
+            {label: "Simple Array", value: 'data/simplearray.json'},
+            {label: "Basic JSON Schema Type", value: 'data/types.json'},
+            {label: 'Basic Radios', value: 'data/radio.json'},
+            {label: 'Condition', value: 'data/condition.json'},
+            {label: 'Kitchen Sink', value: 'data/kitchenSink.json'},
+            {label: 'Login', value: 'data/login.json'},
+            {label: 'Date', value: 'data/date.json'},
+            {label: 'Readonly', value: 'data/readonly.json'},
+            {label: 'Array', value: 'data/array.json'},
+            {label: 'Object', value: 'data/object.json'},
+            {label: 'ArraySelect', value: 'data/arrayselect.json'}
         ],
         validationResult: {},
         schema: {},
         form: [],
         model: {},
-        schemaJson: "",
-        formJson: "",
-        selected: ""
-    };
+        schemaJson: '',
+        formJson: '',
+        selected: ''
+    }
 
-    onSelectChange = val => {
-        //console.log("Selected:" + val);
+    setStateDefault = () => {
+        this.setState({
+            model: this.tempModel,
+        })
+    }
+
+    onSelectChange = (val) => {
+        //console.log('Selected:' + val);
         if (!val) {
             this.setState({
-                schemaJson: "",
-                formJson: "",
-                selected: "",
+                schemaJson: '',
+                formJson: '',
+                selected: '',
                 schema: {},
                 model: {},
                 form: []
-            });
-            return;
+            })
+            return
         }
 
-        $.ajax({
-            type: "GET",
-            url: val.value
-        }).done(
-            function(data) {
-                //console.log('done', data);
-                //console.log('data.schema = ', data.schema);
-                //console.log('data.form = ', data.form);
+        fetch(val.value)
+            .then(x => x.json())
+            .then(({form, schema}) => {
                 this.setState({
-                    schemaJson: JSON.stringify(data.schema, undefined, 2),
-                    formJson: JSON.stringify(data.form, undefined, 2),
+                    schemaJson: JSON.stringify(schema, undefined, 2),
+                    formJson: JSON.stringify(form, undefined, 2),
                     selected: val.value,
-                    schema: data.schema,
+                    schema,
                     model: {},
-                    form: data.form
-                });
-            }.bind(this)
-        );
-    };
+                    form,
+                })
+            })
+    }
 
-    onModelChange = (key, val) => {
-        console.log("ExamplePage.onModelChange:", key, val);
-        var newModel = this.state.model;
-        utils.selectOrSet(key, newModel, val, type);
-        this.setState({ model: newModel });
-    };
+    onModelChange(key, val, type) {
+        console.log('ExamplePage.onModelChange:', key, val)
+        let newModel = this.state.model
+        utils.selectOrSet(key, newModel, val, type)
+        this.setState({model: newModel})
+    }
 
     onValidate = () => {
-        console.log("ExamplePage.onValidate is called");
-        let result = utils.validateBySchema(
-            this.state.schema,
-            this.state.model
-        );
-        this.setState({ validationResult: result });
-    };
+        console.log('ExamplePage.onValidate is called')
+        let result = utils.validateBySchema(this.state.schema, this.state.model)
+        this.setState({validationResult: result})
+    }
 
-    onFormChange = val => {
+    onFormChange = (val) => {
         try {
-            let f = JSON.parse(val);
-            this.setState({ formJson: val, form: f });
-        } catch (e) {}
-    };
+            let f = JSON.parse(val)
+            this.setState({formJson: val, form: f})
+        } catch (e) {
+            console.error(e)
+        }
+    }
 
-    onSchemaChange = val => {
+    onSchemaChange = (val) => {
         try {
-            let s = JSON.parse(val);
-            this.setState({ schemaJson: val, schema: s });
-        } catch (e) {}
-    };
+            let s = JSON.parse(val)
+            this.setState({schemaJson: val, schema: s})
+        } catch (e) {
+            console.error(e)
+        }
+    }
 
     render() {
-        var mapper = {};
-        var schemaForm = "";
-        var validate = "";
+        let mapper = {
+            // 'rc-select': RcSelect
+        }
+
+        let schemaForm = ''
+        let validate = ''
         if (this.state.form.length > 0) {
             schemaForm = (
-                <SchemaForm
-                    schema={this.state.schema}
-                    form={this.state.form}
-                    model={this.state.model}
-                    onModelChange={this.onModelChange}
-                    mapper={mapper}
-                />
-            );
+                <SchemaForm schema={this.state.schema} form={this.state.form} model={this.state.model}
+                            onModelChange={this.onModelChange.bind(this)} mapper={mapper}/>
+            )
             validate = (
                 <div>
-                    <Button
-                        color="primary"
-                        variant="raised"
-                        onClick={this.onValidate}
-                    >
-                        Validate
-                    </Button>
-                    <pre>
-                        {JSON.stringify(
-                            this.state.validationResult,
-                            undefined,
-                            2,
-                            2
-                        )}
-                    </pre>
+                    <Button variant='raised' color='primary' onClick={this.onValidate.bind(this)}>Validate</Button>
+                    <Button variant='raised' color='primary' onClick={this.setStateDefault.bind(this)}>Throw temp model
+                        in</Button>
+                    <pre>{JSON.stringify(this.state.validationResult, undefined, 2)}</pre>
                 </div>
-            );
+            )
         }
 
         return (
-            <div className="col-md-12">
+            <div className='col-md-12'>
                 <h1>Schema Form Example</h1>
-                <div className="row">
-                    <div className="col-sm-4">
-                        <h3 style={{ display: "inline-block" }}>
-                            The Generated Form
-                        </h3>
+                <div className='row'>
+                    <div className='col-sm-4'>
+                        <h3 style={{display: 'inline-block'}}>The Generated Form</h3>
                         {schemaForm}
                         <h3>Model</h3>
-                        <pre>
-                            {JSON.stringify(this.state.model, undefined, 2, 2)}
-                        </pre>
+                        <pre>{JSON.stringify(this.state.model, undefined, 2)}</pre>
                         {validate}
                     </div>
-                    <div className="col-sm-8">
+                    <div className='col-sm-8'>
                         <h3>Select Example</h3>
-                        <div className="form-group">
+                        <div className='form-group'>
                             <Select
-                                name="selectTest"
+                                name='selectTest'
                                 value={this.state.selected}
                                 options={this.state.tests}
-                                onChange={this.onSelectChange}
-                            />
+                                onChange={this.onSelectChange.bind(this)}>
+                            </Select>
                         </div>
                         <h3>Form</h3>
-                        <AceEditor
-                            mode="json"
-                            theme="github"
-                            height="300px"
-                            width="800px"
-                            onChange={this.onFormChange}
-                            name="aceForm"
-                            value={this.state.formJson}
-                            editorProps={{ $blockScrolling: true }}
-                        />
+                        <AceEditor mode='json' theme='github' height='300px' width='800px'
+                                   onChange={this.onFormChange.bind(this)} name='aceForm' value={this.state.formJson}
+                                   editorProps={{$blockScrolling: true}}/>
                         <h3>Schema</h3>
-                        <AceEditor
-                            mode="json"
-                            theme="github"
-                            height="300px"
-                            width="800px"
-                            onChange={this.onSchemaChange}
-                            name="aceSchema"
-                            value={this.state.schemaJson}
-                            editorProps={{ $blockScrolling: true }}
-                        />
+                        <AceEditor mode='json' theme='github' height='300px' width='800px'
+                                   onChange={this.onSchemaChange.bind(this)} name='aceSchema'
+                                   value={this.state.schemaJson} editorProps={{$blockScrolling: true}}/>
                     </div>
                 </div>
             </div>
-        );
+        )
     }
 }
 
-export default ExamplePage;
+module.exports = ExamplePage
