@@ -12,13 +12,7 @@ var _react2 = _interopRequireDefault(_react);
 
 var _styles = require('@material-ui/core/styles');
 
-var _Button = require('@material-ui/core/Button');
-
-var _Button2 = _interopRequireDefault(_Button);
-
-var _IconButton = require('@material-ui/core/IconButton');
-
-var _IconButton2 = _interopRequireDefault(_IconButton);
+var _core = require('@material-ui/core');
 
 var _Delete = require('@material-ui/icons/Delete');
 
@@ -38,20 +32,11 @@ var _ComposedComponent2 = _interopRequireDefault(_ComposedComponent);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-(function () {
-    var enterModule = require('react-hot-loader').enterModule;
-
-    enterModule && enterModule(module);
-})();
-
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
 function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
 
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; } /**
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                * Created by steve on 11/09/15.
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                */
-
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
 
 var styles = function styles(theme) {
     return {
@@ -95,37 +80,16 @@ var Array = function (_React$Component) {
 
         var _this = _possibleConstructorReturn(this, (Array.__proto__ || Object.getPrototypeOf(Array)).call(this, props));
 
-        _this.onAppend = _this.onAppend.bind(_this);
-        _this.onDelete = _this.onDelete.bind(_this);
-        // we have the model here for the entire form, get the model for this array only
-        // and add to the state. if is empty, add an entry by calling onAppend directly.
-        _this.state = {
-            model: _utils2.default.selectOrSet(_this.props.form.key, _this.props.model) || []
-        };
-        //console.log('constructor', this.props.form.key, this.props.model, this.state.model);
-        return _this;
-    }
-
-    _createClass(Array, [{
-        key: 'componentDidMount',
-        value: function componentDidMount() {
-            // Always start with one empty form unless configured otherwise.
-            if (this.props.form.startEmpty !== true && this.state.model.length === 0) {
-                this.onAppend();
-            }
-        }
-    }, {
-        key: 'onAppend',
-        value: function onAppend() {
+        _this.onAppend = function () {
             //console.log('onAppend is called this.state.model', this.state.model);
-            var empty;
-            if (this.props.form && this.props.form.schema && this.props.form.schema.items) {
-                var items = this.props.form.schema.items;
+            var empty = void 0;
+            if (_this.props.form && _this.props.form.schema && _this.props.form.schema.items) {
+                var items = _this.props.form.schema.items;
                 if (items.type && items.type.indexOf('object') !== -1) {
                     empty = {};
 
                     // Check for possible defaults
-                    if (!this.props.options || this.props.options.setSchemaDefaults !== false) {
+                    if (!_this.props.options || _this.props.options.setSchemaDefaults !== false) {
                         empty = typeof items['default'] !== 'undefined' ? items['default'] : empty;
 
                         // Check for defaults further down in the schema.
@@ -141,57 +105,71 @@ var Array = function (_React$Component) {
                     }
                 } else if (items.type && items.type.indexOf('array') !== -1) {
                     empty = [];
-                    if (!this.props.options || this.props.options.setSchemaDefaults !== false) {
+                    if (!_this.props.options || _this.props.options.setSchemaDefaults !== false) {
                         empty = items['default'] || empty;
                     }
                 } else {
                     // No type? could still have defaults.
-                    if (!this.props.options || this.props.options.setSchemaDefaults !== false) {
+                    if (!_this.props.options || _this.props.options.setSchemaDefaults !== false) {
                         empty = items['default'] || empty;
                     }
                 }
             }
-            var newModel = this.state.model;
+            var newModel = _this.state.model;
             Array.assignItemId(empty);
             newModel.push(empty);
-            this.setState({
-                model: newModel
-            });
-            this.props.onChangeValidate(this.state.model);
+            _this.setState({ model: newModel });
+            _this.props.onChangeValidate(_this.state.model);
             //console.log('After append this.state.model', newModel);
-        }
-    }, {
-        key: 'onDelete',
-        value: function onDelete(index) {
-            // console.log('onDelete is called', index);
-            var newModel = this.state.model;
-            newModel.splice(index, 1);
-            this.setState({
-                model: newModel
-            });
-            this.props.onChangeValidate(this.state.model);
-        }
-    }, {
-        key: 'setIndex',
-        value: function setIndex(index) {
+        };
+
+        _this.onDelete = function (index) {
+            return function () {
+                var newModel = _this.state.model;
+                newModel.splice(index, 1);
+                _this.setState({ model: newModel });
+                _this.props.onChangeValidate(_this.state.model);
+            };
+        };
+
+        _this.setIndex = function (index) {
             return function (form) {
                 if (form.key) {
                     form.key[form.key.indexOf('')] = index;
                 }
             };
-        }
-    }, {
-        key: 'copyWithIndex',
-        value: function copyWithIndex(form, index) {
+        };
+
+        _this.copyWithIndex = function (form, index) {
             var copy = _lodash2.default.cloneDeep(form);
             copy.arrayIndex = index;
-            _utils2.default.traverseForm(copy, this.setIndex(index));
+            _utils2.default.traverseForm(copy, _this.setIndex(index));
             return copy;
+        };
+
+        _this.onAppend = _this.onAppend.bind(_this);
+        _this.onDelete = _this.onDelete.bind(_this);
+        // we have the model here for the entire form, get the model for this array only
+        // and add to the state. if is empty, add an entry by calling onAppend directly.
+        _this.state = {
+            model: _utils2.default.selectOrSet(_this.props.form.key, _this.props.model) || []
+            //console.log('constructor', this.props.form.key, this.props.model, this.state.model);
+        };return _this;
+    }
+
+    _createClass(Array, [{
+        key: 'componentDidMount',
+        value: function componentDidMount() {
+            // Always start with one empty form unless configured otherwise.
+            if (this.props.form.startEmpty !== true && this.state.model.length === 0) {
+                this.onAppend();
+            }
         }
     }, {
         key: 'render',
         value: function render() {
-            //console.log('Array.render', this.props.form.items, this.props.model, this.state.model);
+            var _this2 = this;
+
             var _props = this.props,
                 classes = _props.classes,
                 form = _props.form;
@@ -199,24 +177,27 @@ var Array = function (_React$Component) {
             var arrays = [];
             var model = this.state.model;
             // console.log('Array.render', model);
-            for (var i = 0; i < model.length; i++) {
+
+            var _loop = function _loop(i) {
                 var item = model[i];
-                var onItemDelete = this.onDelete.bind(this, i);
                 var forms = form.items.map(function (form, index) {
-                    var copy = this.copyWithIndex(form, i);
-                    return this.props.builder(copy, this.props.model, index, this.props.mapper, this.props.onChange, this.props.builder);
-                }.bind(this));
+                    return _this2.props.builder(_this2.copyWithIndex(form, i), _this2.props.model, index, _this2.props.mapper, _this2.props.onChange, _this2.props.builder);
+                });
                 //console.log('forms', i, forms);
                 arrays.push(_react2.default.createElement(
                     'div',
                     { className: classes.arrayItem, key: item && item[Array.ITEM_ID] || i },
                     _react2.default.createElement(
-                        _IconButton2.default,
-                        { onClick: onItemDelete, className: classes.deleteItemButton },
+                        _core.IconButton,
+                        { onClick: _this2.onDelete(i), className: classes.deleteItemButton },
                         _react2.default.createElement(_Delete2.default, { fontSize: 'small' })
                     ),
                     forms
                 ));
+            };
+
+            for (var i = 0; i < model.length; i++) {
+                _loop(i);
             }
             return _react2.default.createElement(
                 'div',
@@ -236,18 +217,11 @@ var Array = function (_React$Component) {
                     )
                 ),
                 _react2.default.createElement(
-                    _Button2.default,
-                    { className: classes.addButton, variant: 'contained', color: 'primary', onClick: this.onAppend },
+                    _core.Button,
+                    { className: classes.addButton, letiant: 'contained', color: 'primary', onClick: this.onAppend },
                     this.props.form.add || 'Add'
                 )
             );
-        }
-    }, {
-        key: '__reactstandin__regenerateByEval',
-        // @ts-ignore
-        value: function __reactstandin__regenerateByEval(key, code) {
-            // @ts-ignore
-            this[key] = eval(code);
         }
     }], [{
         key: 'getDerivedStateFromProps',
@@ -269,25 +243,4 @@ var Array = function (_React$Component) {
 
 Array.ITEM_ID = '_SCHEMAFORM_ITEM_ID';
 Array._SEQUENCE = 1;
-
-var _default = (0, _ComposedComponent2.default)((0, _styles.withStyles)(styles)(Array));
-
-exports.default = _default;
-;
-
-(function () {
-    var reactHotLoader = require('react-hot-loader').default;
-
-    var leaveModule = require('react-hot-loader').leaveModule;
-
-    if (!reactHotLoader) {
-        return;
-    }
-
-    reactHotLoader.register(styles, 'styles', 'src/Array.js');
-    reactHotLoader.register(Array, 'Array', 'src/Array.js');
-    reactHotLoader.register(_default, 'default', 'src/Array.js');
-    leaveModule(module);
-})();
-
-;
+exports.default = (0, _ComposedComponent2.default)((0, _styles.withStyles)(styles)(Array));
