@@ -28,12 +28,6 @@ var _FormControlLabel2 = _interopRequireDefault(_FormControlLabel);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-(function () {
-    var enterModule = require('react-hot-loader').enterModule;
-
-    enterModule && enterModule(module);
-})();
-
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
 function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
@@ -82,35 +76,9 @@ var FormCheckbox = function (_Component) {
                 })
             );
         }
-    }, {
-        key: '__reactstandin__regenerateByEval',
-        // @ts-ignore
-        value: function __reactstandin__regenerateByEval(key, code) {
-            // @ts-ignore
-            this[key] = eval(code);
-        }
     }]);
 
     return FormCheckbox;
 }(_react.Component);
 
-var _default = (0, _ComposedComponent2.default)(FormCheckbox);
-
-exports.default = _default;
-;
-
-(function () {
-    var reactHotLoader = require('react-hot-loader').default;
-
-    var leaveModule = require('react-hot-loader').leaveModule;
-
-    if (!reactHotLoader) {
-        return;
-    }
-
-    reactHotLoader.register(FormCheckbox, 'FormCheckbox', 'src/Checkbox.js');
-    reactHotLoader.register(_default, 'default', 'src/Checkbox.js');
-    leaveModule(module);
-})();
-
-;
+exports.default = (0, _ComposedComponent2.default)(FormCheckbox);
