@@ -14,12 +14,6 @@ var _react2 = _interopRequireDefault(_react);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-(function () {
-    var enterModule = require('react-hot-loader').enterModule;
-
-    enterModule && enterModule(module);
-})();
-
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
 function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
@@ -29,7 +23,7 @@ function _inherits(subClass, superClass) { if (typeof superClass !== "function" 
 
 var utils = require('./utils');
 
-var _default = function _default(ComposedComponent) {
+exports.default = function (ComposedComponent) {
     var defaultProps = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : {};
     return function (_React$Component) {
         _inherits(_class, _React$Component);
@@ -69,8 +63,7 @@ var _default = function _default(ComposedComponent) {
 
         }, {
             key: 'onChangeValidate',
-            value: function onChangeValidate(e) {
-                // console.log('onChangeValidate e', e);
+            value: function onChangeValidate(e, v) {
                 var value = null;
                 switch (this.props.form.schema.type) {
                     case 'integer':
@@ -88,14 +81,20 @@ var _default = function _default(ComposedComponent) {
                     case 'boolean':
                         value = e.target.checked;
                         break;
-                    case 'object':
-                    case 'date':
+                    case 'tBoolean':
+                        if (e.target.value != 'yes' || e.target.value != 'no') {
+                            value = v;
+                        }
+                        break;
+
                     case 'array':
                         value = e;
                         break;
+                    case 'object':
                     default:
                         value = e.target.value;
                 }
+
                 //console.log('onChangeValidate this.props.form, value', this.props.form, value);
                 var validationResult = utils.validate(this.props.form, value);
                 this.setState({
@@ -103,7 +102,6 @@ var _default = function _default(ComposedComponent) {
                     valid: validationResult.valid,
                     error: validationResult.valid ? null : validationResult.error.message
                 });
-                //console.log('conhangeValidate this.props.form.key, value', this.props.form.key, value);
 
                 this.props.onChange(this.props.form.key, value);
             }
@@ -115,7 +113,7 @@ var _default = function _default(ComposedComponent) {
                 // console.log('Text.defaultValue key', this.props.form.key);
                 // console.log('Text.defaultValue model', this.props.model);
                 var value = void 0;
-                if (props.form.key) value = utils.selectOrSet(props.form.key, props.model);
+                if (props.form && props.form.key) value = utils.selectOrSet(props.form.key, props.model);
                 //console.log('Text defaultValue value = ', value);
 
                 // check if there is a default value
@@ -132,7 +130,6 @@ var _default = function _default(ComposedComponent) {
                 if (!value && props.form.titleMap && props.form.titleMap[0].value) {
                     value = props.form.titleMap[0].value;
                 }
-                //console.log('value', value);
                 return value;
             }
         }, {
@@ -140,33 +137,8 @@ var _default = function _default(ComposedComponent) {
             value: function render() {
                 return _react2.default.createElement(ComposedComponent, _extends({}, defaultProps, this.props, this.state, { onChangeValidate: this.onChangeValidate }));
             }
-        }, {
-            key: '__reactstandin__regenerateByEval',
-            // @ts-ignore
-            value: function __reactstandin__regenerateByEval(key, code) {
-                // @ts-ignore
-                this[key] = eval(code);
-            }
         }]);
 
         return _class;
     }(_react2.default.Component);
 };
-
-exports.default = _default;
-;
-
-(function () {
-    var reactHotLoader = require('react-hot-loader').default;
-
-    var leaveModule = require('react-hot-loader').leaveModule;
-
-    if (!reactHotLoader) {
-        return;
-    }
-
-    reactHotLoader.register(_default, 'default', 'src/ComposedComponent.js');
-    leaveModule(module);
-})();
-
-;
