@@ -18,12 +18,6 @@ var _styles = require('@material-ui/core/styles');
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-(function () {
-    var enterModule = require('react-hot-loader').enterModule;
-
-    enterModule && enterModule(module);
-})();
-
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
 function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
@@ -73,36 +67,9 @@ var FieldSet = function (_React$Component) {
                 )
             );
         }
-    }, {
-        key: '__reactstandin__regenerateByEval',
-        // @ts-ignore
-        value: function __reactstandin__regenerateByEval(key, code) {
-            // @ts-ignore
-            this[key] = eval(code);
-        }
     }]);
 
     return FieldSet;
 }(_react2.default.Component);
 
-var _default = (0, _styles.withStyles)(styles)(FieldSet);
-
-exports.default = _default;
-;
-
-(function () {
-    var reactHotLoader = require('react-hot-loader').default;
-
-    var leaveModule = require('react-hot-loader').leaveModule;
-
-    if (!reactHotLoader) {
-        return;
-    }
-
-    reactHotLoader.register(styles, 'styles', 'src/FieldSet.js');
-    reactHotLoader.register(FieldSet, 'FieldSet', 'src/FieldSet.js');
-    reactHotLoader.register(_default, 'default', 'src/FieldSet.js');
-    leaveModule(module);
-})();
-
-;
+exports.default = (0, _styles.withStyles)(styles)(FieldSet);
