@@ -38,12 +38,6 @@ var _ComposedComponent2 = _interopRequireDefault(_ComposedComponent);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-(function () {
-    var enterModule = require('react-hot-loader').enterModule;
-
-    enterModule && enterModule(module);
-})();
-
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
 function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
@@ -61,8 +55,8 @@ var styles = function styles(theme) {
     };
 };
 
-var Radios = function (_React$Component) {
-    _inherits(Radios, _React$Component);
+var Radios = function (_Component) {
+    _inherits(Radios, _Component);
 
     function Radios() {
         _classCallCheck(this, Radios);
@@ -107,36 +101,9 @@ var Radios = function (_React$Component) {
                 )
             );
         }
-    }, {
-        key: '__reactstandin__regenerateByEval',
-        // @ts-ignore
-        value: function __reactstandin__regenerateByEval(key, code) {
-            // @ts-ignore
-            this[key] = eval(code);
-        }
     }]);
 
     return Radios;
-}(_react2.default.Component);
+}(_react.Component);
 
-var _default = (0, _ComposedComponent2.default)((0, _styles.withStyles)(styles)(Radios));
-
-exports.default = _default;
-;
-
-(function () {
-    var reactHotLoader = require('react-hot-loader').default;
-
-    var leaveModule = require('react-hot-loader').leaveModule;
-
-    if (!reactHotLoader) {
-        return;
-    }
-
-    reactHotLoader.register(styles, 'styles', 'src/Radios.js');
-    reactHotLoader.register(Radios, 'Radios', 'src/Radios.js');
-    reactHotLoader.register(_default, 'default', 'src/Radios.js');
-    leaveModule(module);
-})();
-
-;
+exports.default = (0, _ComposedComponent2.default)((0, _styles.withStyles)(styles)(Radios));
