@@ -54,26 +54,21 @@ var _FieldSet = require('./FieldSet');
 
 var _FieldSet2 = _interopRequireDefault(_FieldSet);
 
+var _TripleBoolean = require('./TripleBoolean');
+
+var _TripleBoolean2 = _interopRequireDefault(_TripleBoolean);
+
 var _lodash = require('lodash');
 
 var _lodash2 = _interopRequireDefault(_lodash);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-(function () {
-    var enterModule = require('react-hot-loader').enterModule;
-
-    enterModule && enterModule(module);
-})();
-
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
 function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
 
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; } /**
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                * Created by steve on 11/09/15.
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                */
-
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
 
 var SchemaForm = function (_React$Component) {
     _inherits(SchemaForm, _React$Component);
@@ -95,6 +90,7 @@ var SchemaForm = function (_React$Component) {
             'checkbox': _Checkbox2.default,
             'help': _Help2.default,
             'array': _Array2.default,
+            'tBoolean': _TripleBoolean2.default,
             'fieldset': _FieldSet2.default
         };
 
@@ -105,24 +101,20 @@ var SchemaForm = function (_React$Component) {
             if (_lodash2.default.isNil(currentValue) && !_lodash2.default.isNil(value)) _this.props.onModelChange(key, value, form.type, form);
         };
 
-        _this.onChange = _this.onChange.bind(_this);
-        _this.builder = _this.builder.bind(_this);
-        return _this;
-    }
+        _this.setDefault = function (key, model, form, value) {
+            var currentValue = _utils2.default.selectOrSet(key, model);
 
-    // Assign default values and save it to the model
+            // If current value is not setted and exist a default, apply the default over the model
+            if (_lodash2.default.isNil(currentValue) && !_lodash2.default.isNil(value)) _this.props.onModelChange(key, value, form.type, form);
+        };
 
-
-    _createClass(SchemaForm, [{
-        key: 'onChange',
-        value: function onChange(key, val) {
+        _this.onChange = function (key, val) {
             //console.log('SchemaForm.onChange', key, val);
-            this.props.onModelChange(key, val);
-        }
-    }, {
-        key: 'builder',
-        value: function builder(form, model, index, mapper, onChange, _builder) {
-            var Field = this.mapper[form.type];
+            _this.props.onModelChange(key, val);
+        };
+
+        _this.builder = function (form, model, index, mapper, onChange, builder) {
+            var Field = _this.mapper[form.type];
             if (!Field) {
                 return null;
             }
@@ -134,7 +126,7 @@ var SchemaForm = function (_React$Component) {
 
             var key = form.key && form.key.join('.') || index;
 
-            var errors = this.props && this.props.errors ? this.props.errors : {};
+            var errors = _this.props && _this.props.errors ? _this.props.errors : {};
             var error = key in errors ? errors[key] : null;
 
             return _react2.default.createElement(Field, {
@@ -142,13 +134,23 @@ var SchemaForm = function (_React$Component) {
                 form: form,
                 key: key,
                 onChange: onChange,
-                setDefault: this.setDefault,
+                setDefault: _this.setDefault,
                 mapper: mapper,
-                builder: _builder,
+                builder: builder,
                 errorText: error
             });
-        }
-    }, {
+        };
+
+        return _this;
+    }
+
+    // Assign default values and save it to the model
+
+
+    // Assign default values and save it to the model
+
+
+    _createClass(SchemaForm, [{
         key: 'render',
         value: function render() {
             var _this2 = this;
@@ -168,32 +170,9 @@ var SchemaForm = function (_React$Component) {
                 forms
             );
         }
-    }, {
-        key: '__reactstandin__regenerateByEval',
-        // @ts-ignore
-        value: function __reactstandin__regenerateByEval(key, code) {
-            // @ts-ignore
-            this[key] = eval(code);
-        }
     }]);
 
     return SchemaForm;
 }(_react2.default.Component);
 
 module.exports = SchemaForm;
-;
-
-(function () {
-    var reactHotLoader = require('react-hot-loader').default;
-
-    var leaveModule = require('react-hot-loader').leaveModule;
-
-    if (!reactHotLoader) {
-        return;
-    }
-
-    reactHotLoader.register(SchemaForm, 'SchemaForm', 'src/SchemaForm.js');
-    leaveModule(module);
-})();
-
-;
