@@ -14,29 +14,11 @@ var _ComposedComponent = require('./ComposedComponent');
 
 var _ComposedComponent2 = _interopRequireDefault(_ComposedComponent);
 
-var _MenuItem = require('@material-ui/core/MenuItem');
+var _core = require('@material-ui/core');
 
-var _MenuItem2 = _interopRequireDefault(_MenuItem);
-
-var _Select = require('@material-ui/core/Select');
-
-var _Select2 = _interopRequireDefault(_Select);
-
-var _InputLabel = require('@material-ui/core/InputLabel');
-
-var _InputLabel2 = _interopRequireDefault(_InputLabel);
-
-var _FormControl = require('@material-ui/core/FormControl');
-
-var _FormControl2 = _interopRequireDefault(_FormControl);
+var _lodash = require('lodash');
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-(function () {
-    var enterModule = require('react-hot-loader').enterModule;
-
-    enterModule && enterModule(module);
-})();
 
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
@@ -67,16 +49,28 @@ var Select = function (_Component) {
 
         var _this = _possibleConstructorReturn(this, (Select.__proto__ || Object.getPrototypeOf(Select)).call(this, props));
 
-        _this.onSelected = function (event) {
-            var currentValue = event.target.value;
-            _this.setState({
-                currentValue: currentValue
-            });
+        _this.onSelected = function (_ref2) {
+            var currentValue = _ref2.target.value;
+
+            _this.setState({ currentValue: currentValue });
             _this.props.onChangeValidate(event);
         };
 
+        _this.onSelected = _this.onSelected.bind(_this);
+
+        var _this$props = _this.props,
+            model = _this$props.model,
+            form = _this$props.form;
+        var key = form.key;
+
+
+        var storedValue = model && _this.getModelKey(model, key) || false;
+        var defaultValue = form.schema.default || false;
+        var value = !(0, _lodash.isEmpty)(storedValue) && storedValue || defaultValue;
+
+        _this.props.setDefault(key, model, form, value);
         _this.state = {
-            currentValue: getModelValue(_this.props.model, _this.props.form)
+            currentValue: value
         };
         return _this;
     }
@@ -88,37 +82,29 @@ var Select = function (_Component) {
 
             var menuItems = form.titleMap.map(function (item, idx) {
                 return _react2.default.createElement(
-                    _MenuItem2.default,
+                    _core.MenuItem,
                     { key: idx, value: item.value },
                     item.name
                 );
             });
             return _react2.default.createElement(
-                _FormControl2.default,
-                { fullWidth: true },
+                _core.FormControl,
+                { style: { width: '100%' } },
                 _react2.default.createElement(
-                    _InputLabel2.default,
-                    null,
-                    form.title
+                    _core.InputLabel,
+                    { htmlFor: 'age-simple' },
+                    this.props.form.title
                 ),
                 _react2.default.createElement(
-                    _Select2.default,
+                    _core.MuiSelect,
                     {
-                        value: this.state.currentValue,
+                        disabled: this.props.form.readonly,
                         placeholder: form.title,
-                        disabled: form.readonly,
-                        onChange: this.onSelected
-                    },
+                        value: this.state.currentValue || '',
+                        onChange: this.onSelected },
                     menuItems
                 )
             );
-        }
-    }, {
-        key: '__reactstandin__regenerateByEval',
-        // @ts-ignore
-        value: function __reactstandin__regenerateByEval(key, code) {
-            // @ts-ignore
-            this[key] = eval(code);
         }
     }], [{
         key: 'getDerivedStateFromProps',
@@ -134,24 +120,4 @@ var Select = function (_Component) {
     return Select;
 }(_react.Component);
 
-var _default = (0, _ComposedComponent2.default)(Select);
-
-exports.default = _default;
-;
-
-(function () {
-    var reactHotLoader = require('react-hot-loader').default;
-
-    var leaveModule = require('react-hot-loader').leaveModule;
-
-    if (!reactHotLoader) {
-        return;
-    }
-
-    reactHotLoader.register(getModelValue, 'getModelValue', 'src/Select.js');
-    reactHotLoader.register(Select, 'Select', 'src/Select.js');
-    reactHotLoader.register(_default, 'default', 'src/Select.js');
-    leaveModule(module);
-})();
-
-;
+exports.default = (0, _ComposedComponent2.default)(Select);
