@@ -14,17 +14,9 @@ var _ComposedComponent = require('./ComposedComponent');
 
 var _ComposedComponent2 = _interopRequireDefault(_ComposedComponent);
 
-var _TextField = require('@material-ui/core/TextField');
-
-var _TextField2 = _interopRequireDefault(_TextField);
+var _core = require('@material-ui/core');
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-(function () {
-    var enterModule = require('react-hot-loader').enterModule;
-
-    enterModule && enterModule(module);
-})();
 
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
@@ -32,8 +24,8 @@ function _possibleConstructorReturn(self, call) { if (!self) { throw new Referen
 
 function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
 
-var Text = function (_React$Component) {
-    _inherits(Text, _React$Component);
+var Text = function (_Component) {
+    _inherits(Text, _Component);
 
     function Text(props) {
         _classCallCheck(this, Text);
@@ -60,7 +52,7 @@ var Text = function (_React$Component) {
                 error = _props.error,
                 onChangeValidate = _props.onChangeValidate;
 
-            return _react2.default.createElement(_TextField2.default, {
+            return _react2.default.createElement(_core.TextField, {
                 type: form.type,
                 label: form.title,
                 placeholder: form.placeholder,
@@ -72,35 +64,9 @@ var Text = function (_React$Component) {
                 fullWidth: true
             });
         }
-    }, {
-        key: '__reactstandin__regenerateByEval',
-        // @ts-ignore
-        value: function __reactstandin__regenerateByEval(key, code) {
-            // @ts-ignore
-            this[key] = eval(code);
-        }
     }]);
 
     return Text;
-}(_react2.default.Component);
+}(_react.Component);
 
-var _default = (0, _ComposedComponent2.default)(Text);
-
-exports.default = _default;
-;
-
-(function () {
-    var reactHotLoader = require('react-hot-loader').default;
-
-    var leaveModule = require('react-hot-loader').leaveModule;
-
-    if (!reactHotLoader) {
-        return;
-    }
-
-    reactHotLoader.register(Text, 'Text', 'src/Text.js');
-    reactHotLoader.register(_default, 'default', 'src/Text.js');
-    leaveModule(module);
-})();
-
-;
+exports.default = (0, _ComposedComponent2.default)(Text);
