@@ -14,17 +14,11 @@ var _ComposedComponent = require('./ComposedComponent');
 
 var _ComposedComponent2 = _interopRequireDefault(_ComposedComponent);
 
-var _TextField = require('@material-ui/core/TextField');
+var _core = require('@material-ui/core');
 
-var _TextField2 = _interopRequireDefault(_TextField);
+var _utils = require('./utils');
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-(function () {
-    var enterModule = require('react-hot-loader').enterModule;
-
-    enterModule && enterModule(module);
-})();
 
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
@@ -50,7 +44,8 @@ var TextArea = function (_React$Component) {
                 error = _props.error,
                 onChangeValidate = _props.onChangeValidate;
 
-            return _react2.default.createElement(_TextField2.default, {
+            value = (0, _utils.selectOrSet)(this.props.form.key, this.props.model) || '';
+            return _react2.default.createElement(_core.TextField, {
                 type: form.type,
                 label: form.title,
                 placeholder: form.placeholder,
@@ -65,35 +60,9 @@ var TextArea = function (_React$Component) {
                 fullWidth: true
             });
         }
-    }, {
-        key: '__reactstandin__regenerateByEval',
-        // @ts-ignore
-        value: function __reactstandin__regenerateByEval(key, code) {
-            // @ts-ignore
-            this[key] = eval(code);
-        }
     }]);
 
     return TextArea;
 }(_react2.default.Component);
 
-var _default = (0, _ComposedComponent2.default)(TextArea);
-
-exports.default = _default;
-;
-
-(function () {
-    var reactHotLoader = require('react-hot-loader').default;
-
-    var leaveModule = require('react-hot-loader').leaveModule;
-
-    if (!reactHotLoader) {
-        return;
-    }
-
-    reactHotLoader.register(TextArea, 'TextArea', 'src/TextArea.js');
-    reactHotLoader.register(_default, 'default', 'src/TextArea.js');
-    leaveModule(module);
-})();
-
-;
+exports.default = (0, _ComposedComponent2.default)(TextArea);
