@@ -20,12 +20,6 @@ var _AutoComplete2 = _interopRequireDefault(_AutoComplete);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-(function () {
-    var enterModule = require('react-hot-loader').enterModule;
-
-    enterModule && enterModule(module);
-})();
-
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
 function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
@@ -116,36 +110,9 @@ var TextSuggest = function (_React$Component) {
                 })
             );
         }
-    }, {
-        key: '__reactstandin__regenerateByEval',
-        // @ts-ignore
-        value: function __reactstandin__regenerateByEval(key, code) {
-            // @ts-ignore
-            this[key] = eval(code);
-        }
     }]);
 
     return TextSuggest;
 }(_react2.default.Component);
 
-var _default = (0, _ComposedComponent2.default)(TextSuggest);
-
-exports.default = _default;
-;
-
-(function () {
-    var reactHotLoader = require('react-hot-loader').default;
-
-    var leaveModule = require('react-hot-loader').leaveModule;
-
-    if (!reactHotLoader) {
-        return;
-    }
-
-    reactHotLoader.register(dataSourceConfig, 'dataSourceConfig', 'src/TextSuggest.js');
-    reactHotLoader.register(TextSuggest, 'TextSuggest', 'src/TextSuggest.js');
-    reactHotLoader.register(_default, 'default', 'src/TextSuggest.js');
-    leaveModule(module);
-})();
-
-;
+exports.default = (0, _ComposedComponent2.default)(TextSuggest);
