@@ -1,19 +1,26 @@
 'use strict';
 
-jest.dontMock('../SchemaForm');
-jest.dontMock('../utils');
-jest.dontMock('lodash');
+var _react = require('react');
 
-var React = require('react');
-var ShallowRenderer = require('react-test-renderer/shallow');
-var SchemaForm = require('../SchemaForm');
+var _react2 = _interopRequireDefault(_react);
 
-describe('SchemaForm', function () {
+var _SchemaForm = require('../SchemaForm');
 
-  beforeEach(function () {});
+var _SchemaForm2 = _interopRequireDefault(_SchemaForm);
+
+var _enzyme = require('enzyme');
+
+var _enzymeAdapterReact = require('enzyme-adapter-react-16');
+
+var _enzymeAdapterReact2 = _interopRequireDefault(_enzymeAdapterReact);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+(0, _enzyme.configure)({ adapter: new _enzymeAdapterReact2.default() });
+
+describe('SchemaForm test', function () {
 
   it('shows SchemaForm', function () {
-    var shallowRenderer = new ShallowRenderer();
     var cfg = {
       form: {},
       schema: {
@@ -22,13 +29,12 @@ describe('SchemaForm', function () {
       model: {},
       mapper: {}
     };
-    shallowRenderer.render(React.createElement(SchemaForm, {
+
+    var result = (0, _enzyme.shallow)(_react2.default.createElement(_SchemaForm2.default, {
       schema: cfg.schema,
       mapper: cfg.mapper
     }));
 
-    var result = shallowRenderer.getRenderOutput();
-    expect(result.type).toEqual('div');
-    expect(result.props.children).toEqual([]);
+    expect(result.props().className).toEqual('SchemaForm');
   });
 });
